Make Metadata and BadgeGroup common
